% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automagic_tabs.R
\name{automagic_tabs}
\alias{automagic_tabs}
\title{Generacion automatica de tabs}
\usage{
automagic_tabs(input_data, panel_name, .output, .layout = NULL, ...)
}
\arguments{
\item{input_data}{tibble con al menos 2 columnas, una para el título de las tabs y otra con la salida a mostrar.}

\item{panel_name}{string con el nombre de la columna de ID.}

\item{.output}{string con el nombre de la columna del output.}

\item{.layout}{string que representa el layout de las tabs, puede tomar los valores "l-body" ,"l-body-outset",
"l-page" y "l-screen". Por defecto el valor es NULL y toma como parámetro "l-body".}

\item{...}{parámetros adicionales que corresponden a todos los disponibles en los chunks de
rmarkdown (fig.align, fig.width, …)}
}
\value{
string que contiene concatenados todos los chunks generados automáticamente.
}
\description{
Permite generar automáticamente el código necesario para agrupar múltiples chunks de Rmarkdown en tabs.
Concatenando todos los chunks en un string que puede ser posteriormente tejido y renderizado.
}
\details{
dado un tiblle, que debe contener una columna de “ID” (representará al título de las tabs) y otra columna
que almacena la salida que se desee generar (gráfico, texto, código, ...), se genera automáticamente un
string que posteriormente puede ser renderizado en un documento Rmarkdown.
}
\examples{

library(dplyr)
library(ggplot2)
library(purrr)

dataset <-
 tibble::tibble(ID = c("A","B","C","D"),
                numbers = list(rnorm(10), runif(10), rpois(10,2), rexp(10))) \%>\%
 mutate(plots = map(numbers,
                    ~ data.frame(plots = .x) \%>\% ggplot(aes(x=plots)) + geom_histogram()))

dataset

automagic_tabs(input_data = dataset, panel_name = "ID", .output = "plots")

unlink("figure", recursive = TRUE)

}
\seealso{
\href{https://rafzamb.github.io/sknifedatar/}{sknifedatar website}
}
