% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_avellaneda.R
\docType{data}
\name{data_avellaneda}
\alias{data_avellaneda}
\title{Flujo vehicular por peaje Avellaneda en la Ciudad Autónoma de Buenos Aires.}
\format{
Un data frame con 4383 filas y 2 columnas, la variable date representa la fecha diaria, value la cantidad de vehículos que pasaron ese día por ese peaje.
}
\source{
\url{https://data.buenosaires.gob.ar/dataset/flujo-vehicular-por-unidades-peaje-ausa}
}
\usage{
data_avellaneda
}
\description{
Los datos corresponden al \href{https://data.buenosaires.gob.ar/dataset/flujo-vehicular-por-unidades-peaje-ausa}(Flujo de vehículos livianos por Peaje Avellaneda en la Ciudad Autónoma de Buenos Aires).
Los datos disponibles incluyen el período Enero 2009 a Diciembre 2020.
}
\keyword{datasets}
