% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercepcion_calles.R
\docType{data}
\name{intercepcion_calles}
\alias{intercepcion_calles}
\title{Muestra la intercepciòn de las principales calles y avenidas de la ciudad de Buenos Aires}
\format{
A data frame con 2417 filas y 3 columnas:
\describe{
\item{id}{id de la esquina}
\item{lat}{latitud}
\item{long}{longitud}
...
}
}
\source{
\url{https://www.openstreetmap.org/}
}
\usage{
intercepcion_calles
}
\description{
Conjunto de datos que registra la fecha, franja horaria, tipo de delito y geolocalización de crimenes ocurridos entre el 2017 y 2019
diamonds.
}
\keyword{datasets}
