% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listN.R
\name{listN}
\alias{listN}
\title{Agregaci칩n autom치tica del nombre de elementos de una lista}
\usage{
listN(...)
}
\arguments{
\item{...}{}
}
\value{

}
\description{
Agregaci칩n autom치tica del nombre de elementos de una lista
}
