% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multirefit.R
\name{modeltime_multirefit}
\alias{modeltime_multirefit}
\title{Función para reajustar el o los modelos para múltiples series de tiempo}
\usage{
modeltime_multirefit(models_table)
}
\arguments{
\item{models_table}{tibble que proviene del objeto "\strong{table_time}" de la salida de la función "\strong{modeltime_multifit}".}
}
\value{
Devuelve el objeto "\strong{table_time}" recalibrado.
}
\description{
Esta función permite aplicar la función "\strong{modeltime_refit()}" de \href{https://business-science.github.io/modeltime/}{\strong{modeltime}}
a múltiples series y modelos.
}
\details{
Toma como input el objeto "\strong{table_time}" de la salida de la función "\strong{modeltime_multifit}" y
devuelve el mismo objeto pero entrenado en todo el período.
}
\examples{

# Data
library(modeltime)
data_serie <- sknifedatar::table_time
table_time <- data_serie$table_time

# modeltime_multirefit
table_time_refit <- sknifedatar::modeltime_multirefit(models_table = table_time)

table_time_refit
}
