% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_wfs_bestmodel.R
\name{modeltime_wfs_bestmodel}
\alias{modeltime_wfs_bestmodel}
\title{Modeltime best workflow from a set of models}
\usage{
modeltime_wfs_bestmodel(
  .wfs_results,
  .model = NULL,
  .metric = "rmse",
  .minimize = TRUE
)
}
\arguments{
\item{.wfs_results}{a tibble generated from the modeltime_wfs_fit function}

\item{.model}{model to select best model from}

\item{.metric}{metric to get best model from ('mae', 'mape','mase','smape','rmse','rsq')}

\item{.minimize}{a boolean indicating whether to minimize (TRUE) or maximize (FALSE) the metric}
}
\value{
a tibble containing the best model based on the selected metric
}
\description{
Get best combination of model and recipe based on a a wffits object generated from the modeltime_wfs_fit function
}
\details{
The best model is selected based on a specific metric ('mae', 'mape','mase','smape','rmse','rsq'). The default is to minimize the metric. However, if the model is being selected based on rsq minimize should be FALSE.
}
\examples{
library(sknifedatar)
library(recipes)
library(modeltime)
library(workflowsets)
library(workflows)
library(parsnip)
library(timetk)

data <- sknifedatar::data_avellaneda \%>\% 
mutate(date=as.Date(date)) \%>\% 
  filter(date<'2012-06-01')

recipe_date <- recipe(value ~ ., data = data) \%>\% 
  step_date(date, features = c('dow','doy','week','month','year')) 

recipe_date_lag <- recipe_date \%>\% 
  step_lag(value, lag = 7) \%>\% 
  step_ts_impute(all_numeric(), period=365)

mars <- mars(mode = 'regression') \%>\%
  set_engine('earth')

wfsets <- workflow_set(
  preproc = list(
    R_date = recipe_date,
    R_date_lag = recipe_date_lag),
  models  = list(M_mars = mars),
  cross   = TRUE)

wffits <- modeltime_wfs_fit(.wfsets = wfsets, 
                            .split_prop = 0.8, 
                            .serie=data)

wfbest <- modeltime_wfs_bestmodel(.wfs_results = wffits,
                      .metric='rsq',
                      .minimize = FALSE)
wfbest
}
