% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_wfs_multirefit.R
\name{modeltime_wfs_multirefit}
\alias{modeltime_wfs_multirefit}
\title{Refit one or more trained workflows to new data}
\usage{
modeltime_wfs_multirefit(models_table)
}
\arguments{
\item{models_table}{a tibble that comes from the output of the \code{modeltime_wfs_multifit()}, \code{modeltime_wfs_multiforecast()},
\code{modeltime_wfs_multibestmodel()} functions. For the \code{modeltime_wfs_multifit} function,
the 'table_time' object must be selected from the output.}
}
\value{
a tibble, corresponds to the same tibble supplied in the 'models_table' parameter but with the refit
of the workflows saved in the 'nested_model' column.
}
\description{
It allows retraining a set of workflows trained on new data.
}
\examples{
library(dplyr)
library(earth)

df <- sknifedatar::emae_series

datex <- '2020-02-01'
df_emae <- df \%>\% 
  dplyr::filter(date <= datex) \%>\% 
  tidyr::nest(nested_column=-sector) \%>\% 
  head(2)

receta_base <- recipes::recipe(value ~ ., data = df \%>\% select(-sector))

mars <- parsnip::mars(mode = 'regression') \%>\% parsnip::set_engine('earth')

wfsets <- workflowsets::workflow_set(
  preproc = list(
    R_date = receta_base),
  models  = list(M_mars = mars),
  cross   = TRUE)

wfsets_fit <- modeltime_wfs_multifit(.wfs = wfsets,
                                     .prop = 0.8, 
                                     serie = df_emae)

sknifedatar::modeltime_wfs_multirefit(wfsets_fit$table_time)

}
