% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_wfs_rank.R
\name{modeltime_wfs_rank}
\alias{modeltime_wfs_rank}
\title{Modeltime workflow sets ranking based on a metric}
\usage{
modeltime_wfs_rank(.wfs_results, rank_metric = NULL, minimize = TRUE)
}
\arguments{
\item{.wfs_results}{a tibble generated with the modeltime_wfs_fit function}

\item{rank_metric}{the metric used to generate the ranking 'mae', 'mape','mase','smape','rmse','rsq'.}

\item{minimize}{a boolean indicating whether to minimize (TRUE) or maximize (FALSE) the metric}
}
\value{
A tibble containing the models ranked by a specific metric
}
\description{
This function generates a ranking of models generated with modeltime_wfs_fit
}
\details{
The ranking depends on the metric selected
}
\examples{
library(sknifedatar)
library(recipes)
library(modeltime)
library(workflowsets)
library(workflows)
library(parsnip)
library(timetk)

data <- sknifedatar::data_avellaneda \%>\% 
  mutate(date=as.Date(date)) \%>\% 
  filter(date<'2012-06-01')

recipe_date <- recipe(value ~ ., data = data) \%>\% 
  step_date(date, features = c('dow','doy','week','month','year')) 

recipe_date_lag <- recipe_date \%>\% 
  step_lag(value, lag = 7) \%>\% 
  step_ts_impute(all_numeric(), period=365)

mars <- mars(mode = 'regression') \%>\%
  set_engine('earth')

wfsets <- workflow_set(
  preproc = list(
    R_date = recipe_date,
    R_date_lag = recipe_date_lag),
  models  = list(M_mars = mars),
  cross   = TRUE)

wffits <- modeltime_wfs_fit(.wfsets = wfsets, 
                            .split_prop = 0.8, 
                            .serie=data)
wffits

modeltime_wfs_rank(.wfs_results = wffits,
                   rank_metric = 'rsq',
                   minimize = FALSE)
}
\seealso{
\href{https://rafzamb.github.io/sknifedatar/}{sknifedatar website}
}
