% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multifit.R
\name{multi_fit}
\alias{multi_fit}
\title{Ajuste de multiples modelos a multiples series de tiempo}
\usage{
multi_fit(serie, .prop, ...)
}
\arguments{
\item{serie}{series de tiempo anidadas}

\item{.prop}{proporción de partición de las series en train/test}

\item{...}{modelos o workflows a entrenar (model_1,model2,...)}
}
\value{
una lista con 2 elementos. El primer componnete es un tibble que contiene una primera columna que contiene el nombre
de las series, seguidamente la comuna "nested_column" que almacena las series de tiempo, luego una columna para cada
modelo suministrado donde se almacenanan los modelos o workflows entrenados para cada series. Finalmente las columnas
nested_model y calibration que guardan los n modelos entrenados para cada serie y las metricas de ajuste sobre la
particion de test.
}
\description{
Esta funcion permite ajusatr multiples mdoelos sobre multiples series de tiempo,
utilizando los modelos del paquete \href{https://business-science.github.io/modeltime/}{modeltime}.
}
\details{
El enfoque de esta función no esta relacionado con series de panel, esta orientada a multiples series individuales.
Recibeindo como primer argumento "serie" un conjunto de series anidadas (por ejemplo a traves de la funcion nest),
luego especificando una proporcion deseada para particionar cada una de las series en train/test.\preformatted{     Por ultimo se deben suministrar los modelos a entrenar, siemplente escribiendo el nombre de los modelos separados
     por comas, la funcion admite tantos modelos como se requierean.
}

@importFrom rlang .data

@seealso \href{https://rafzamb.github.io/sknifedatar/}{sknifedatar website}
}
