% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pertenencia_punto.R
\name{pertenencia_punto}
\alias{pertenencia_punto}
\title{Pertenencia de localizaciones}
\usage{
pertenencia_punto(data, referencia, metros)
}
\arguments{
\item{data}{: dataframe con las columnas de geolocalización nombradas como "long" y "lat"}

\item{referencia}{: dataframe con las columnas de geolocalización nombradas como "long" y "lat"}

\item{metros}{: distancia en metros}
}
\value{

}
\description{
Pertenencia de localizaciones
}
