% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/automagic_tabs.R
\name{automagic_tabs}
\alias{automagic_tabs}
\title{Automatic Generation of Tabs}
\usage{
automagic_tabs(input_data, panel_name, .output, .layout = NULL, ...)
}
\arguments{
\item{input_data}{tibble with at least 2 columns, one for the title of the tabs and another with
the output to be displayed.}

\item{panel_name}{string with the name of the ID column.}

\item{.output}{string with the name of the column of the output.}

\item{.layout}{string that represents the layout of the tabs, can take the values "l-body",
"l-body-outset", "l-page" and "l-screen". By default the value is NULL and takes
"l-body" as parameter.}

\item{...}{additional parameters that correspond to all those available in rmarkdown chunks
(fig.align, fig.width, ...).}
}
\value{
concatenated string of all automatically generated chunks.
}
\description{
It allows to automatically generate the code necessary to group multiple Rmarkdown chunks
into tabs. Concatenating all the chunks into a string that can be later knitted and rendered.
}
\details{
given a tibble, which must contain an "ID" column (representing the title of the tabs) and another
column that stores the output to be generated (plot, text, code, ...), a string is automatically
generated which can be later rendered in a Rmarkdown document.
}
\examples{

library(dplyr)
library(ggplot2)
library(purrr)
library(sknifedatar)

dataset <-
 tibble::tibble(ID = c("A","B"),
                numbers = list(rnorm(5), runif(5))) \%>\%
 mutate(plots = map(numbers,
                    ~ data.frame(plots = .x) \%>\% ggplot(aes(x=plots)) + geom_histogram()))

dataset

automagic_tabs(input_data = dataset, panel_name = "ID", .output = "plots")

unlink("figure", recursive = TRUE)

}
\seealso{
\href{https://rafzamb.github.io/sknifedatar/}{sknifedatar website}
}
