% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_avellaneda.R
\docType{data}
\name{data_avellaneda}
\alias{data_avellaneda}
\title{Vehicle flow through Avellaneda toll in Ciudad Autónoma de Buenos Aires, Argentina.}
\format{
A dataframe with 4383 rows y 2 columns: date is the daily date, value is the number of vehicles no that day.
}
\source{
\url{https://data.buenosaires.gob.ar/dataset/flujo-vehicular-por-unidades-peaje-ausa}
}
\usage{
data_avellaneda
}
\description{
Data corresponds to \href{https://data.buenosaires.gob.ar/dataset/flujo-vehicular-por-unidades-peaje-ausa}{Vehicle flow through Avellaneda toll in Ciudad Autónoma de Buenos Aires.}.
From January 2009 to December 2020.
}
\keyword{datasets}
