% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_na.R
\name{insert_na}
\alias{insert_na}
\title{Function para agregar valores NA a un data frame}
\usage{
insert_na(.dataset, columns, .p = 0.01, seed = 123)
}
\arguments{
\item{.dataset}{data frame}

\item{columns}{vector que indica el nombre de las columnas donde se agregaran los valores NA, ejemplo de formato c("X1","X2").}

\item{.p}{valor entre 0 y 1, indica la proporción de valores NA que se agregaran.}

\item{seed}{semilla de números aleatorios.}
}
\value{
El data frame original suministrado a data, pero con los valores NA agregados en las columnas indicadas.
}
\description{
Esta función permite agregar valores NA a un data frame, pudiendo seleccionar las columnas
y la propòrcion de NAs deseados.
}
\examples{
insert_na(.dataset = iris, columns = c("Sepal.Length","Petal.Length"), .p = 0.25)
}
