% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlapply.R
\name{mlapply}
\alias{mlapply}
\title{Title}
\usage{
mlapply(.Fun, ..., .Cluster = NULL, .parFun = parallel::parLapply)
}
\arguments{
\item{.Fun}{asADSa}

\item{...}{asas}

\item{.Cluster}{dasd}

\item{.parFun}{dasasda}
}
\value{
saSa
}
\description{
Title
}
