% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multibestmodel.R
\name{modeltime_multibestmodel}
\alias{modeltime_multibestmodel}
\title{Function to obtain the best model from a modeltime table}
\usage{
modeltime_multibestmodel(.table, .metric = NULL, .minimize = TRUE)
}
\arguments{
\item{.table}{"\strong{table_time}" tibble generated with the "\strong{modeltime_multifit}" function.}

\item{.metric}{evaluation metric, from "\strong{modeltime_accuracy}": 'mae', 'mape','mase','smape','rmse','rsq'.}

\item{.minimize}{boolean (default = TREUE), TRUE if the error metric should be minimized, FALSE in order to maximize it.}
}
\value{
table_time tibble filtered by the best model.
}
\description{
This feature allows you to select the best model for each series, based on a specific evaluation metric.
}
\details{
This function takes the object "** table_time \strong{" from the output of the function "} modeltime_multifit **",
and selects the best model based on the selected metric.
}
\examples{

## libraries
library(modeltime)
library(sknifedatar)

# Data
data_serie <- sknifedatar::table_time
                                      
# best_model_emae
best_model_emae <- modeltime_multibestmodel(.table = data_serie$table_time,
                                            .metric = "rmse",
                                            .minimize = TRUE)

best_model_emae

}
