% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multibestmodel.R
\name{modeltime_multibestmodel}
\alias{modeltime_multibestmodel}
\title{Función para obtener el mejor modelo a partir de una modeltime table}
\usage{
modeltime_multibestmodel(.table, .metric = "mase", .optimization = which.min)
}
\arguments{
\item{.table}{tibble que proviene del objeto table_time de la salida de la funcion "modeltime_multifit".}

\item{.metric}{métrica de evaluación, proviene de modeltime_accuracy: 'mae', 'mape','mase','smape','rmse','rsq'}

\item{.optimization}{Es posible minimizar una métrica de error ('which.min') o maximizar el rsq ('which.max')}
}
\value{
tibble table_time filtrada por el mejor modelo
}
\description{
Esta funcion permite seleccionar el mejor modelo para cada serie, en función de determinada métrica de evaluación
}
\details{
Esta funcion toma el objeto table_time de la salida de la funcion "modeltime_multifit", posteriomente se selecciona el mejor modelo en función de la métrica seleccionada
}
\examples{

library(sknifedatar)
data("table_time")

best_model_emae <- modeltime_multibestmodel(.table=table_time,
                                            .metric=rmse,
                                            .optimization = which.min)

best_model_emae
}
