% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multifit.R
\name{modeltime_multifit}
\alias{modeltime_multifit}
\title{Fit Multiple Models to Multiple Time Series}
\usage{
modeltime_multifit(serie, .prop, ...)
}
\arguments{
\item{serie}{nested time series.}

\item{.prop}{series train/test partition ratio.}

\item{...}{models or workflows to train (model_1, model2, ...).}
}
\value{
A list of 2 items. The first component is a tibble with a first column that contains the
name of the series, and a second column called "nested_column" that stores the time series,
then a column for each model where the trained models or workflows for each series are stored.
The last 2 columns, "nested_model" and "calibration", store the "n" trained models for each
series and the adjustment metrics on the test partition.
}
\description{
This function allows multiple models to be adjusted over multiple time series, using models
from the \href{https://business-science.github.io/modeltime/}{modeltime} package.
}
\details{
The focus of this function is not related to panel series, it is oriented to multiple individual
series. Receiving as the first argument "series" a set of nested series (for example through the
nest function), then specifying a desired train/test partition ratio for series. The
final input to the function are the models to be trained, simply by typing the name
of the models separated by commas. The function admits as many models as required.
}
\examples{

# Generate "table_time" object
## libraries
library(modeltime)
library(parsnip)

## Data
nested_serie <- 
tidyr::nest(dplyr::filter(sknifedatar::emae_series, date < '2006-02-01'),
            nested_column = -sector)

## Models
m_ets <- set_engine(exp_smoothing(), 'ets')

m_nnetar <- set_engine(nnetar_reg(), "nnetar")

# modeltime_multifit
sknifedatar::modeltime_multifit(serie = head(nested_serie,2),
                                .prop = 0.97,
                                m_ets,
                                m_nnetar)

}
\seealso{
\href{https://rafzamb.github.io/sknifedatar/}{sknifedatar website}
}
