% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multifit.R
\name{modeltime_multifit}
\alias{modeltime_multifit}
\title{Ajuste de multiples modelos a multiples series de tiempo}
\usage{
modeltime_multifit(serie, .prop, ...)
}
\arguments{
\item{serie}{series de tiempo anidadas}

\item{.prop}{proporción de partición de las series en train/test}

\item{...}{modelos o workflows a entrenar (model_1,model2,...)}
}
\value{
una lista con 2 elementos. El primer componnete es un tibble que contiene una primera columna que contiene el nombre
de las series, seguidamente la comuna "nested_column" que almacena las series de tiempo, luego una columna para cada
modelo suministrado donde se almacenanan los modelos o workflows entrenados para cada series.\preformatted{    Finalmente las columnas "nested_model" y "calibration" que guardan los "n" modelos entrenados para cada serie y las metricas de ajuste sobre la
    particion de test.
}
}
\description{
Esta funcion permite ajusatr multiples mdoelos sobre multiples series de tiempo,
utilizando los modelos del paquete \href{https://business-science.github.io/modeltime/}{modeltime}.
}
\details{
El enfoque de esta función no esta relacionado con series de panel, esta orientada a multiples series individuales.
Recibeindo como primer argumento "serie" un conjunto de series anidadas (por ejemplo a traves de la funcion nest),
luego especificando una proporcion deseada para particionar cada una de las series en train/test.\preformatted{     Por ultimo se deben suministrar los modelos a entrenar, siemplente escribiendo el nombre de los modelos separados
     por comas, la funcion admite tantos modelos como se requierean.
}
}
\examples{
# libraries
library(modeltime)
library(rsample)
library(parsnip)
library(recipes)
library(workflows)
library(dplyr)
library(tidyr)
library(sknifedatar)

# Data
data("emae_series")
nested_serie = emae_series \%>\% nest(nested_column=-sector)

# Recipes
recipe_1 = recipe(value ~ ., data = emae_series \%>\% select(-sector)) \%>\%
step_date(date, features = c("month", "quarter", "year"), ordinal = TRUE)

# Models
m_auto_arima <- arima_reg() \%>\% set_engine('auto_arima')

m_stlm_arima <- seasonal_reg() \%>\%
  set_engine("stlm_arima")

m_nnetar <- workflow() \%>\%
  add_recipe(recipe_1) \%>\%
  add_model(nnetar_reg() \%>\% set_engine("nnetar"))

# modeltime_multifit
model_table_emae = modeltime_multifit(serie = nested_serie \%>\% head(3),
                                      .prop = 0.8,
                                      m_auto_arima,
                                      m_stlm_arima,
                                      m_nnetar)

model_table_emae
}
\seealso{
\href{https://rafzamb.github.io/sknifedatar/}{sknifedatar website}
}
