% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multiforecast.R
\name{modeltime_multiforecast}
\alias{modeltime_multiforecast}
\title{Forecasting de multiples modelos sobre multiples series}
\usage{
modeltime_multiforecast(models_table, .h = NULL, .prop = NULL)
}
\arguments{
\item{models_table}{tibble que proviene del objeto table_time de la salida de la funcion "modeltime_multifit".}

\item{.h}{horizonte de predicción de la funcion "modeltime_forecast".}

\item{.prop}{proporción de la partición de splits de als series de tiempo. En caso de no especificar "h", esta funcion
proyecta en la particion de testing.}
}
\value{
tibble correspondiente al mismmo suministardo en el parametro "models_table", agregando una nueva columna llamada
"nested_forecast" donde se guardan las predicciones.
}
\description{
Esta funcion permite ralizar un forecating sobre multiples series de tiempo a partir de multiples modelos
entrenados.
}
\details{
Esta funcion toma el objeto table_time de la salida de la funcion "modeltime_multifit", posteriomente
aplica la funcion "modeltime_forecast" a cada uno de los multiples modelos sobre multiples series.
}
\examples{
library(sknifedatar)
data("table_time")

forecast_emae <- modeltime_multiforecast(table_time,
                                         .prop=0.8)


forecast_emae
}
