% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multirefit.R
\name{modeltime_multirefit}
\alias{modeltime_multirefit}
\title{Función para re ajustar el o los modelos para múltiples series de tiempo}
\usage{
modeltime_multirefit(models_table)
}
\arguments{
\item{models_table}{tibble que proviene del objeto table_time de la salida de la funcion "modeltime_multifit".}
}
\value{
Devuelve el objeto table_time recalibrado
}
\description{
Esta función permite aplicar la función modeltime_refit() de modeltime a múltiples series y modelos
}
\details{
Toma como input el objeto table_time de la salida de la funcion "modeltime_multifit" y devuelve el mismo objeto pero entrenado en todo el período
}
\examples{

library(sknifedatar)
library(modeltime)

data("table_time")

table_time_refit <- modeltime_multirefit(models_table = table_time)
table_time_refit
}
