% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multirefit.R
\name{modeltime_multirefit}
\alias{modeltime_multirefit}
\title{Función para reajustar el o los modelos para múltiples series de tiempo}
\usage{
modeltime_multirefit(models_table)
}
\arguments{
\item{models_table}{tibble que proviene del objeto "\strong{table_time}" de la salida de la función "\strong{modeltime_multifit}".}
}
\value{
Devuelve el objeto "\strong{table_time}" recalibrado.
}
\description{
Esta función permite aplicar la función "\strong{modeltime_refit()}" de \href{https://business-science.github.io/modeltime/}{\strong{modeltime}}
a múltiples series y modelos.
}
\details{
Toma como input el objeto "\strong{table_time}" de la salida de la función "\strong{modeltime_multifit}" y
devuelve el mismo objeto pero entrenado en todo el período.
}
\examples{

# Generate "table_time" object
## libraries
library(modeltime)
library(rsample)
library(parsnip)
library(recipes)
library(workflows)
library(dplyr)
library(tidyr)
library(sknifedatar)

## Data
emae_series <- sknifedatar::emae_series
nested_serie <- emae_series \%>\% filter(date < '2008-02-01') \%>\% nest(nested_column=-sector)

## Recipes
recipe_1 = recipe(value ~ ., data = emae_series \%>\% select(-sector)) \%>\%
  step_date(date, features = c("month", "quarter", "year"), ordinal = TRUE)

## Models
m_naive_reg <- naive_reg() \%>\%  set_engine("naive")

m_nnetar <- workflow() \%>\%
  add_model(nnetar_reg() \%>\% set_engine("nnetar")) \%>\%
  add_recipe(recipe_1)

## modeltime_multifit
model_table_emae <- modeltime_multifit(serie = nested_serie \%>\% head(2),
                                       .prop = 0.8,
                                       m_naive_reg ,
                                       m_nnetar)

table_time <- model_table_emae$table_time

# modeltime_multirefit
table_time_refit <- modeltime_multirefit(models_table = table_time)

table_time_refit
}
