% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_multirefit.R
\name{modeltime_multirefit}
\alias{modeltime_multirefit}
\title{Refit the model or models for multiple time series}
\usage{
modeltime_multirefit(models_table)
}
\arguments{
\item{models_table}{'table_time' tibble generated from the \code{modeltime_multifit()} function.}
}
\value{
retrained 'table_time' object.
}
\description{
applies the \code{modeltime_refit()} function from the 'modeltime' package to multiple series and models.
}
\details{
it takes the 'table_time' tibble generated with the  \code{modeltime_multifit()} function and
returns the same object but with the models fitted for the complete period.
}
\examples{
# Data
library(modeltime)
data_serie <- sknifedatar::table_time
table_time <- data_serie$table_time

# modeltime_multirefit
sknifedatar::modeltime_multirefit(models_table = table_time)

}
