% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_wfs_forecast.R
\name{modeltime_wfs_forecast}
\alias{modeltime_wfs_forecast}
\title{Modeltime workflow sets forecast}
\usage{
modeltime_wfs_forecast(.wfs_results, .series, .split_prop = NULL, .h = NULL)
}
\arguments{
\item{.wfs_results}{tibble of combination of recipes and models fitted, generated with the modeltime_wfs_fit function}

\item{.series}{time series dataframe}

\item{.split_prop}{time series split proportion}

\item{.h}{time series horizon from the modeltime_forecast() function}
}
\value{
a tibble containing the forecast for each model
}
\description{
Forecast from a set of recipes and models trained by modeltime_wfs_fit
}
\details{
Since it uses the modeltime_forecast() function from modeltime, either the forecast can be made on new data or on a number of periods.
}
\examples{
library(sknifedatar)
library(recipes)
library(modeltime)
library(workflowsets)
library(workflows)
library(parsnip)
library(timetk)
data <- sknifedatar::data_avellaneda \%>\% 
mutate(date=as.Date(date)) \%>\% 
  filter(date<'2012-01-01')

recipe_date <- recipe(value ~ ., data = data) \%>\% 
  step_date(date, features = c('week','month', 'quarter', 'semester', 'year')) 

recipe_date_lag <- recipe_date \%>\% 
  step_lag(value, lag = 1) \%>\% 
  step_ts_impute(all_numeric(), period=365)

mars <- mars(mode = 'regression') \%>\%
  set_engine('earth')

wfsets <- workflow_set(
  preproc = list(
    R_date = recipe_date,
    R_date_lag = recipe_date_lag),
  models  = list(M_mars = mars),
  cross   = TRUE)

wffits <- modeltime_wfs_fit(.wfsets = wfsets, 
                            .split_prop = 0.8, 
                            .serie=data)
wfforecast <- modeltime_wfs_forecast(.wfs_results=wffits, 
                                     .series = data,
                                     .split_prop = 0.8) 
wfforecast
}
