% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeltime_wfs_heatmap.R
\name{modeltime_wfs_heatmap}
\alias{modeltime_wfs_heatmap}
\title{Modeltime workflowsets heatmap plot}
\usage{
modeltime_wfs_heatmap(
  .wfs_results,
  metric = "rsq",
  low_color = "#c7e9b4",
  high_color = "#253494"
)
}
\arguments{
\item{.wfs_results}{a tibble generated with the modeltime_wfs_fit function}

\item{metric}{a metric the metric used for the heatmap values: 'mae', 'mape','mase','smape','rmse','rsq'.}

\item{low_color}{color for the worst metric (highest error or lowest rsq)}

\item{high_color}{color for the better metric (lowest error or highest rsq)}
}
\value{
a ggplot heatmap
}
\description{
Function to generate a heatmap for each recipe and model on a wffits object generated with the modeltime_wfs_fit function
}
\details{
This function assumes that the models included in the wffits object are named M_{name_of_model}, since the .model_id is {recipe_name}\emph{M}{name_of_model} and the '\emph{M}' is used to separate the recipe from the model name.
}
\examples{
library(sknifedatar)
library(recipes)
library(modeltime)
library(workflowsets)
library(workflows)
library(parsnip)
library(timetk)

data <- sknifedatar::data_avellaneda \%>\% 
  mutate(date=as.Date(date)) \%>\% 
  filter(date<'2011-01-01')

recipe_date <- recipe(value ~ ., data = data) \%>\% 
  step_date(date, features = c('dow','doy','week','month','year')) 

recipe_date_fourier <- recipe_date \%>\%
  step_fourier(date, period = 365, K=1)

mars_backward <- mars(prune_method='backward',
                      mode = 'regression') \%>\%
  set_engine('earth')

mars_forward <- mars(prune_method = 'forward', 
                     mode='regression') \%>\% 
  set_engine('earth')

wfsets <- workflow_set(
  preproc = list(
    date = recipe_date,
    fourier = recipe_date_fourier),
  models  = list(M_mars_backward = mars_backward, 
                 M_mars_forward = mars_forward),
  cross   = TRUE)

wffits <- modeltime_wfs_fit(.wfsets = wfsets, 
                            .split_prop = 0.6, 
                            .serie=data)

modeltime_wfs_heatmap(wffits, 'rsq')
}
