% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multieval.R
\name{multieval}
\alias{multieval}
\title{Evaluación de múltiples métricas y predicciones}
\usage{
multieval(.dataset, .observed, .predictions, .metrics, value_table = FALSE)
}
\arguments{
\item{.dataset}{data frame con las predicciones, debe tener como mínimo la columna con los datos observados y al menos una columna que haga referencia a las predicciones de un modelo.}

\item{.observed}{string con el nombre de la columna que contiene los datos observados.}

\item{.predictions}{string o vector de strings las columnas donde se almacenan las predicicones.}

\item{.metrics}{métrica o conjunto de métricas que se desean evaluar, las métricas hacen referencias a las permitidas por
el paquete \href{https://yardstick.tidymodels.org/}{yardstick} de tidymodels.}

\item{value_table}{TRUE para mostrar las métricas desagregadas.}
}
\value{
data frame con 4 columnas, siendo las mismas las métricas de evaluación, el estimador usado, el valor de la métrica y el nombre del modelo.
}
\description{
Para un conjunto de predicciones de distintos modelos, permite evaluar múltiples métricas y devolver los resultados en un formato tabular que facilita la comparación de las predicciones.
}
\examples{
set.seed(123)
library(yardstick) # métricas
library(erer) # función listn

predictions <-
  data.frame(truth = runif(100),
             predict_model_1 = rnorm(100, mean = 1,sd =2),
             predict_model_2 = rnorm(100, mean = 0,sd =2),
             predict_model_3 = rnorm(100, mean = 0,sd =3))

multieval(.dataset = predictions,
          .observed = "truth",
          .predictions = c("predict_model_1","predict_model_2","predict_model_3"),
          .metrics = listn(rmse, rsq, mae),
          value_table = TRUE)

# Output ----------------------
# A tibble: 9 x 4
# .metric .estimator .estimate model
# <chr>   <chr>          <dbl> <chr>
#   1 mae     standard     1.45    predict_model_1
# 2 mae     standard     1.67    predict_model_2
# 3 mae     standard     2.43    predict_model_3
# 4 rmse    standard     1.78    predict_model_1
# 5 rmse    standard     2.11    predict_model_2
# 6 rmse    standard     3.01    predict_model_3
# 7 rsq     standard     0.00203 predict_model_1
# 8 rsq     standard     0.0158  predict_model_2
# 9 rsq     standard     0.00254 predict_model_3

#$summary_table
# A tibble: 3 x 4
#  model             mae  rmse     rsq
#  <chr>           <dbl> <dbl>   <dbl>
#  1 predict_model_1  1.45  1.78 0.00203
#  2 predict_model_2  1.67  2.11 0.0158
#  3 predict_model_3  2.43  3.01 0.00254
}
\seealso{
\href{https://rafael-zambrano-blog-ds.netlify.app/posts/2020-12-22-prediccin-de-delitos-en-caba/#modelos-de-series-de-tiempo}{Predicción de delitos/multieval}
}
