% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sliding_window.R
\name{sliding_window}
\alias{sliding_window}
\title{Implemnetacion de sliding window}
\usage{
sliding_window(data, inicio, pliegues, variables)
}
\arguments{
\item{data}{dataframe}

\item{inicio}{vector}

\item{pliegues}{vector}

\item{variables}{vector}
}
\value{
A list
}
\description{
Implemnetacion de sliding window
}
